+++ Hearthstone Deck Tracker/IGameHandler.cs
@@ -122,6 +122,7 @@ namespace Hearthstone_Deck_Tracker
 		void HandlePlayerAbyssalCurse(int value);
 		void HandleOpponentAbyssalCurse(int value);
 		void HandleQuestRewardDatabaseId(int id, int value);
+		void HandlePlayerCorpsesLeftChange(int value);
 		void HandleOpponentCorpsesLeftChange(int value);
 	}
 }