+++ Hearthstone Deck Tracker/Windows/OverlayWindow.xaml.cs
@@ -26,6 +26,7 @@ using System.Windows.Input;
 using Hearthstone_Deck_Tracker.Controls.Overlay;
 using Hearthstone_Deck_Tracker.Controls.Overlay.BoardOrder;
 using Hearthstone_Deck_Tracker.Controls.Overlay.Mercenaries;
+using Hearthstone_Deck_Tracker.Hearthstone.CounterSystem.Settings;
 using Hearthstone_Deck_Tracker.Hearthstone.Entities;
 using Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem;
 using Hearthstone_Deck_Tracker.Utility.RemoteData;
@@ -210,6 +211,12 @@ namespace Hearthstone_Deck_Tracker.Windows
 			OverlayExtensions.OnToolTipChanged += SetTooltip;
 
 			_game = game;
+
+			CounterVisibilitySettings.Instance.Changed += (_, _) =>
+			{
+				Core.UpdatePlayerResourcesWidget();
+				Core.UpdateOpponentResourcesWidget();
+			};
 
 			for(int i = 0; i < MaxBoardSize; i++)
 			{
@@ -1149,9 +1156,9 @@ namespace Hearthstone_Deck_Tracker.Windows
 			LinkOpponentDeckDisplay.Hide(true);
 		}
 
-		internal void UpdatePlayerResourcesWidget(int maxHealth, int maxMana, int maxHandSize, int? maxGold)
+		internal void UpdatePlayerResourcesWidget(int maxHealth, int maxMana, int maxHandSize, int? corpsesLeft, int? maxGold)
 		{
-			PlayerResourcesViewModel.UpdatePlayerResourcesWidget(maxHealth, maxMana, maxHandSize, maxGold: maxGold);
+			PlayerResourcesViewModel.UpdatePlayerResourcesWidget(maxHealth, maxMana, maxHandSize, corpsesLeft, maxGold);
 		}
 
 		internal void UpdateOpponentResourcesWidget(int maxHealth, int maxMana, int maxHandSize, int? corpsesLeft)